+++ src/redux/modules/entities/skills/selectors.js
@@ -8,16 +8,17 @@ import { getCurrentForm } from '../../form/selectors';
 import { getDisplay } from '../users/selectors';
 
 export const getHasProficiencyFormValue = state =>
-  getCurrentForm(state) &&
-  getCurrentForm(state).getIn(['values', 'hasProficiency']);
+  getCurrentForm(state) && getCurrentForm(state).getIn(['values', 'hasProficiency']);
 
 export const getSkillMemberSidePanelTableItems = state => {
   const selectedEntityId = getSelectedEntityId(state);
+  const hasProficiency = getHasProficiencyFormValue(state);
   return getDependantEntityTableItems(state).map(user => ({
     ...user,
     name: getDisplay(user),
-    proficiency: user.skills
-      .filter(skill => selectedEntityId === skill.id)
-      .map(skill => skill.proficiency)[0]
+    hasProficiency,
+    proficiency: hasProficiency
+      ? user.skills.filter(skill => selectedEntityId === skill.id).map(skill => skill.proficiency)[0]
+      : null
   }));
 };